extract Text bind logic to TextViewUtils
