+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/CardViewHolder.java
@@ -102,7 +102,7 @@ public final class CardViewHolder extends ParentViewHolder<Card> {
 
     private void bindLabel() {
         final Text label = mModel != null ? mModel.getLabel() : null;
-        Text.bind(label, mLabel, R.dimen.text_size_card_label, Styles.getPrimaryColor(mModel));
+        TextViewUtils.bind(label, mLabel, R.dimen.text_size_card_label, Styles.getPrimaryColor(mModel));
         mDivider.setBackgroundColor(Styles.getTextColor(mModel));
     }
